web/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
websockets==11.0.3
pydantic==2.3.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.20
alembic==1.12.0
psycopg2-binary==2.9.7
aiofiles==23.2.1
python-dotenv==1.0.0
roslibpy==1.5.0
numpy==1.25.2
pyyaml==6.0.1
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1
coverage==7.3.0
black==23.7.0
isort==5.12.0
flake8==6.1.0
mypy==1.5.1
